requirements.txt
amqp==5.0.6
appdirs==1.4.4
asgiref==3.4.1
billiard==3.6.4.0
black==21.7b0
braintree==3.59.0
cairocffi==1.2.0
CairoSVG==2.5.2
celery==5.1.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
cryptography==3.4.7
cssselect2==0.4.1
defusedxml==0.7.1
Django==3.2.5
django-environ==0.4.5
django-parler==2.2
django-rosetta==0.9.6
django-taggit==1.5.1
flower==1.0.0
html5lib==1.1
humanize==3.10.0
idna==3.2
kombu==5.1.0
Markdown==3.3.4
mypy-extensions==0.4.3
oauthlib==3.1.1
pathspec==0.9.0
Pillow==8.3.1
polib==1.1.1
prometheus-client==0.11.0
prompt-toolkit==3.0.19
psycopg2-binary==2.9.1
pycparser==2.20
PyJWT==2.1.0
pyphen==0.11.0
python3-openid==3.2.0
pytz==2021.1
regex==2021.7.6
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
tinycss2==1.1.0
tomli==1.0.4
tornado==6.1
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
WeasyPrint==52.5
webencodings==0.5.1
